v1/serverless-llama/pyproject.toml
[project]
name = "python-uv-starter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "nitric ==1.2.1",
    "betterproto ==2.0.0b6",
    "llama-cpp-python>=0.3.1",
]

[tool.uv]
dev-dependencies = [
    "watchdog>=5.0.3",
]
